test_workflow_env/requirements_no_talib.txt
# Pattern Scanner Requirements
yfinance>=0.2.18
pandas>=1.5.3
numpy>=1.24.3
scipy>=1.10.1
mplfinance>=0.12.8b0
azure-storage-blob>=12.16.0
python-dotenv>=1.0.0
matplotlib>=3.7.1
git+https://github.com/white07S/TradingPatternScanner.git
